Query: Match joined tables properly when lifting for group by aggregate
Earlier we only matched alias and skipped if alias matched. This could happen if the table name starts with same character.
Added logic to unwrap the join table and match exact table name/schema to identify if it is parent table
This may not be the best fix but it avoids different table matching and ends up generating invalid SQL in normal scenario.
The additional tables from navigation are going to be joinExpression only. If we find a shape which is not what we expect we fallback to previous behavior.
The inner table of joinExpression may not always be table (think of a case where there is query filter probably). Though fixing for a complicated case may cause instability in running queries. Deferring for that till customer reports. In most cases running into this bug will generate invalid SQL exception

Resolves #27163